copy over the setLinks code from TaglessII
as previously discovered, in order for updates to work, links need to be (re)swizzled, as new links will come in with keys instead of object references and some nodes may be new and others will disappear.

This is all the stuff that makes it so hard to do updating force layout arrangements in vanilla D3, as it requires some knowledge of the internal implementation.